new programs/qasm.py file
